Migrate formating to separate modules
Migrate to a more pluggable view framework. Now each output format
is specified in its own header and source file. Each output format
must declare a format_t structure with function pointers to handle
the various stages of output generation.

A set of format_new_t/format_free_t functions can be used to allocate
a state object that is unique to that format. This state object is
passed on other calls. Currently the only format to use this is the
json format (used to determine when to print a comma to between each
row).