Ensure decimal values less than 1 start with leading 0
JSON requires that numbers have a leading digit. For numbers less than 1,
IBM i CLI does not place a leading 0, which causes invalid JSON numeric values.

We hack around this by allocating an extra byte for the leading 0 and fix-up
the value if it's found to not include it before we pass it to the output callback.