Add old JSON container format for compatibility
The old format returns an array of objects inside an object with a single
key "records". The whole encasing object is redundant and should be removed,
but for compatibility we will keep it and have it by default.

The container format can be switched using the DB2UTIL_JSON_CONTAINER
environment variable. Setting it to "array" gets the new format, while
setting it to "object" or having it unset will give the old format. It is
recommended to set the environment variable instead of leaving it unset,
as the default may change in a future version.