fix: Save bound pointer for columns
In #12, the strlen() call was moved from the row functions in to
db2util_query. Because numeric data types adjust col->data, this means
we are now running strlen() on a pointer which may have been adjusted
previously and before it is adjusted for the current data. This can
cause the length to be off when passed to the row function and a NUL
byte accidentally output.

Instead, we add a new field to the col_info_t struct to save off the
pointer passed to SQLBindCol. This ensures our strlen() uses the same
pointer that SQLFetch put data in to and our lengths are consistent no
matter what value col->data has.

Fixes #14