fix: Bind input parameter length properly
By passing the precision of the described type instead of the precision
of the actual type we're passing, we can get discrepancies that lead to
unexpected results. In the case of an INTEGER parameter, it will have a
precision of 4 (bytes) but a character length of up to 10. Since we bind
as character, CLI will treat this length as the character string length
(even though it's _supposed_ to use the bind indicator for this value
instead, *sigh*). Since we always bind as character and let CLI do the
conversion, we can eliminate the SQLDescribeParam and just pass the
length of the command line argument as the precision. The SQL type and
scale are not actually used in this case so we can specify "dummy"
values for them.

Fixes #16